+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - added config path override (use JVM argument `-Dzume.configPathOverride="path/to/zume.json5"`)
 - added `thirdPersonToggleMode` config option
 - fixed bounce when deactivating third-person zoom
+- fixed Vintage implementation bug with some mixin loaders
 - easing now uses exponent instead of named methods
 - streamlined release process
 - further improvements to overall system stability and other minor adjustments have been made to enhance the user experience